rename GenericPressable to BaseGenericPressable
